Allow underscores in command names
`plugcli` defaults to using `-` in command names, but allows
users to use either `_` or `-`. Internally, this is done by
changing user input from `_` to `-`. But that requires that command
writers define their commands with `-` as well.

This PR will allow command writers to use either `_` or `-`. We
still normalize on `-`, but now the internally-used name is changed
on plugin registration.